+++ __init__.py
@@ -1,2 +1,2 @@
 import ranger.api
-from .oscyank import oscyank
+from .oscyank import oscyank, set_oscyank